environments/example_terminal_bench_env/pyproject.toml
[project]
name = "example-terminal-bench-env"
description = "RLVR environment for TerminalBench2-style tasks using sandboxed bash interaction and pytest verification."
tags = ["multi-turn", "tool-use", "terminal", "sandbox", "train", "eval"]
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "verifiers>=0.1.10",
    "datasets",
    "pyyaml",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = ["example_terminal_bench_env.py", "pyproject.toml", "sample_tasks/**/*"]

[tool.verifiers.eval]
num_examples = 1
rollouts_per_example = 1
